refactor(android): split foreground geolocation helpers
